+++ rmk/CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.5] - 2025-07-06
+
+### Fixed
+
+- "No" key with tailing whitespace cannot be parsed
+- Key processing error when using tap-hold keys in combo
+
+## [0.7.4] - 2025-07-03
+
+### Added
+
+- Add sleep mode for split central after connected to the host
+
+### Updated
+
+- Refactor key processing, fix tap-hold issues
+- Only the valid macro data is stored in the storage now. **Clearing storage is required to update**
+
+### Fixed
+
+- Light service is wrongly disabled
+- Correctly update connection parameters after connected to the host
+- Remove need for quotes on OSM 
+
 ## [0.7.3] - 2025-06-18
 
 ### Added
@@ -20,8 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed sdc build error
-- Fixed cloud build script for ESP32
+- Fix sdc build error
+- Fix cloud build script for ESP32
 
 ## [0.7.2] - 2025-06-12